Music.Prelude.CmdLine

Description

Provides framework for building simple command-line converter programs such as music2midi, music2pdf etc.

Documentation

converterMain

Arguments

:: String
Name of converter function.

-> String
Extension of generated file.

-> IO ()

Generates a basic converter program such as music2ly, music2musicxml etc.

lilypondConverterMain

Arguments

:: String
A file extension (supported by Lilypond).

-> IO ()

Generates a basic converter program that invokes lilypond to generate music notation. Used for music2pdf etc.

translateFile

Arguments

:: String
Translate function (of type FilePath -> music -> IO ()).

-> String
Output file suffix.

-> Maybe String
Prelude to use.

-> Maybe FilePath
Input file.

-> Maybe FilePath
Output file.

-> IO ()

Translate an input file using the given paramters.

translateFileAndRunLilypond

Arguments

:: String
Output file suffix/format passed to Lilypond.

-> Maybe String
Prelude to use.

-> Maybe FilePath
Input file.

-> IO ()

Translate an input file and invoke Lilypond on the resulting output file.

version :: Version

Current Music Suite version.

versionString :: String

Current Music Suite version as a string.